requirements.txt
flask
bs4
pandas
flask-jsonpify
flask-sqlalchemy
flask_restful
lxml
